update done! Falta fazer para todos os campos e testar erros

+++ PortalNet-BackEnd/src/main/java/com/polarising/PortalNet/Repository/ClientRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.polarising.PortalNet.model.Client;
 
 @Repository
-public interface ClientRepository extends CrudRepository<Client, Long>{
+public interface ClientRepository extends CrudRepository<Client, Integer>{
 	
 	List<Client> findByClientId(int clientId);
 }
